# openapi.yaml
openapi: 3.1.0
info:
  title: Business Management & POS System API
  version: 1.0.0
  summary: REST + JSON contract for the single-business web-based POS and operations platform.
  description: Authoritative machine-readable contract derived from API-Architecture-V1.0.md. Server is the only authority
    for business rules, accounting, and inventory. All derived balances (COGS, payment_state, AR/AP/CRL/SRec, cash_balance,
    inventory_value, moving_average_unit_cost, on_hand_quantity) are server-computed. Posted records are immutable; corrections
    go through lifecycle actions (cancel, return, refund). Cancellation, return, and refund are distinct events. All amounts
    use NUMERIC(15,2); quantities NUMERIC(15,4). Single currency (IDR). All timestamps are ISO 8601 with offset; server normalizes
    to UTC. Session-based auth (Bearer opaque tokens) per §2 of the API Architecture.
  contact:
    name: Business POS System API Team
  license:
    name: Proprietary
servers:
- url: https://api.example.com/api/v1
  description: Production
- url: https://staging-api.example.com/api/v1
  description: Staging
- url: http://localhost:8000/api/v1
  description: Local development
...
paths:
...
  /cost-types:
    get:
      operationId: listCostTypes
      summary: List cost types.
      description: List cost types.
      tags:
      - Cost Types
      security:
      - bearerAuth: []
      x-required-capability: cost_type.view
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Q'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CostType'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  links:
                    $ref: '#/components/schemas/Links'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createCostTypes
      summary: Create cost type.
      description: Create cost type.
      tags:
      - Cost Types
      security:
      - bearerAuth: []
      x-required-capability: cost_type.manage
      parameters:
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostTypeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /cost-types/{id}:
    parameters:
    - name: id
      in: path
      required: true
      description: Cost Type id
      schema:
        type: integer
        format: int64
    get:
      operationId: getCostTypes
      summary: Get cost type.
      description: Get cost type.
      tags:
      - Cost Types
      security:
      - bearerAuth: []
      x-required-capability: cost_type.view
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostType'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      operationId: updateCostTypes
      summary: Update cost type.
      description: Update cost type.
      tags:
      - Cost Types
      security:
      - bearerAuth: []
      x-required-capability: cost_type.manage
      parameters:
      - $ref: '#/components/parameters/IfMatchRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostTypePatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      operationId: deleteCostTypes
      summary: Hard delete cost type (only if unreferenced).
      description: Hard delete cost type (only if unreferenced).
      tags:
      - Cost Types
      security:
      - bearerAuth: []
      x-required-capability: cost_type.manage
      parameters:
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '204':
          description: No Content
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          x-error-codes: referenced_by_history
        '404':
          $ref: '#/components/responses/NotFound'
  /cost-types/{id}/deactivate:
    post:
      operationId: deactivateCostTypes
      summary: Deactivate cost type. Preserves history.
      description: Deactivate cost type. Preserves history.
      tags:
      - Cost Types
      security:
      - bearerAuth: []
      x-required-capability: cost_type.manage
      parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: integer
          format: int64
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeactivateRequest'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /dashboard:
    get:
      operationId: getDashboard
      summary: KPI cards + charts. Requires `finance.view_profit` and `finance.view_cash` (return 403 if missing).
      description: KPI cards + charts. Requires `finance.view_profit` and `finance.view_cash` (return 403 if missing).
      tags:
      - Dashboard
      security:
      - bearerAuth: []
      x-required-capability: finance.view_profit
      parameters:
      - $ref: '#/components/parameters/Period'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/CompareTo'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/inventory:
    get:
      operationId: getDashboardInventory
      summary: Inventory-only KPIs (low-stock count, total value).
      description: Inventory-only KPIs (low-stock count, total value).
      tags:
      - Dashboard
      security:
      - bearerAuth: []
      x-required-capability: inventory.view
      parameters:
      - $ref: '#/components/parameters/Period'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque
      description: 'Session-token auth per API Architecture §2. Server-issued opaque random tokens after credential exchange.
        256-bit random, base64url-encoded, stored hashed in the sessions table. Client does NOT parse them. Send `Authorization:
        Bearer <token>`. 401 = missing/expired/revoked. 403 = authenticated but missing capability.'
  parameters:
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Required. See `IdempotencyKey`.
      schema:
        type: string
        format: uuid
        maxLength: 100
...
    IfMatchRequired:
      name: If-Match
      in: header
      required: true
      description: Required for lifecycle actions. See `IfMatch`.
      schema:
        type: string
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page number.
    PerPage:
      name: per_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
      description: Page size. Hard cap 500.
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: Comma-separated list. Prefix `-` for descending.
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
        maxLength: 200
      description: Case-insensitive substring search.
    From:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: ISO 8601 lower bound.
    To:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: ISO 8601 upper bound.
    Period:
      name: period
      in: query
      required: false
      schema:
        type: string
        enum:
        - today
        - this_week
        - this_month
        - this_year
        - custom
      description: Convenience period alias.
    CompareTo:
      name: compare_to
      in: query
      required: false
      schema:
        type: string
      description: 'Period to compare against: `period` alias or `from,to`.'
...
  responses:
    BadRequest:
      description: Validation failure (400).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, or expired credentials (401).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but missing required capability (403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorizationErrorEnvelope'
    NotFound:
      description: Resource not found (404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: State conflict (409).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
      - error
      description: Canonical error envelope. Used for all 4xx/5xx responses.
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
              description: Stable, machine-readable snake_case code.
              example: not_found
            message:
              type: string
              description: Human-readable English.
            details:
              description: Optional, code-specific structure.
              oneOf:
              - type: object
                additionalProperties: true
              - type: array
                items:
                  type: object
                  additionalProperties: true
              - type: 'null'
            request_id:
              type: string
              format: uuid
              description: Echoes `X-Request-ID` header.
    ValidationErrorEnvelope:
      allOf:
      - $ref: '#/components/schemas/ErrorEnvelope'
      - type: object
        description: Returned for `validation_failed` (400).
        properties:
          error:
            type: object
            properties:
              code:
                const: validation_failed
              details:
                type: array
                items:
                  type: object
                  required:
                  - field
                  - code
                  - message
                  properties:
                    field:
                      type: string
                    code:
                      type: string
                    message:
                      type: string
    AuthorizationErrorEnvelope:
      allOf:
      - $ref: '#/components/schemas/ErrorEnvelope'
      - type: object
        description: Returned for `capability_required` (403).
        properties:
          error:
            type: object
            properties:
              code:
                const: capability_required
              details:
                type: object
                properties:
                  required:
                    type: array
                    items:
                      type: string
                    example:
                    - sale.cancel
    ConflictErrorEnvelope:
      allOf:
      - $ref: '#/components/schemas/ErrorEnvelope'
      - type: object
        description: Returned for state conflicts (409).
        properties:
          error:
            type: object
            properties:
              code:
                type: string
                enum:
                - insufficient_stock
                - negative_stock_disallowed
                - payment_exceeds_total
                - allocation_exceeds_payable
                - refund_exceeds_crl
                - repayment_exceeds_srec
                - cash_insufficient
                - return_quantity_exceeds_original
                - purchase_return_quantity_exceeds_original
                - lifecycle_state_invalid
                - already_cancelled
                - cannot_cancel_after_return
                - production_inputs_exceed_stock
                - production_finished_cost_invalid
                - cost_type_deactivated
                - product_deactivated
                - concurrent_modification
                - serialization_failure
                - idempotency_violation
                - delete_not_permitted
                - referenced_by_history
                - version_mismatch
              details:
                type: object
                additionalProperties: true
    Pagination:
      type: object
      required:
      - page
      - per_page
      - total
      - total_pages
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 500
          example: 50
        total:
          type: integer
          minimum: 0
          example: 1234
        total_pages:
          type: integer
          minimum: 0
          example: 25
    Links:
      type: object
      properties:
        self:
          type: string
          format: uri-reference
        next:
          type:
          - string
          - 'null'
          format: uri-reference
        prev:
          type:
          - string
          - 'null'
          format: uri-reference
...
    CostType:
      type: object
      required:
      - id
      - code
      - name
      - is_active
      - version
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 100
        is_active:
          type: boolean
        version:
          type: integer
        code:
          type: string
          minLength: 1
          maxLength: 50
          pattern: ^[a-z0-9_]+$
    CostTypeRequest:
      type: object
      required:
      - code
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        is_active:
          type: boolean
          default: true
        code:
          type: string
          minLength: 1
          maxLength: 50
          pattern: ^[a-z0-9_]+$
    CostTypePatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        is_active:
          type: boolean
...
    PeriodComparison:
      type: object
      properties:
        previous_period:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
        delta:
          type: object
          additionalProperties:
            type: number
        delta_pct:
          type: object
          additionalProperties:
            type: number
...
    InventoryReportResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            total_inventory_value:
              type: number
            products_count:
              type: integer
            low_stock_count:
              type: integer
            out_of_stock_count:
              type: integer
...
    DashboardResponse:
      type: object
      required:
      - kpis
      - as_of
      properties:
        kpis:
          type: object
          properties:
            total_sales:
              type: number
            total_purchases:
              type: number
            net_profit:
              type:
              - number
              - 'null'
            cash_on_hand:
              type:
              - number
              - 'null'
            inventory_value:
              type: number
            low_stock_count:
              type: integer
        charts:
          type: object
          properties:
            sales_trend:
              type: array
              items:
                type: object
                additionalProperties: true
            best_sellers:
              type: array
              items:
                type: object
                additionalProperties: true
            expense_breakdown:
              type: array
              items:
                type: object
                additionalProperties: true
        comparison:
          $ref: '#/components/schemas/PeriodComparison'
        as_of:
          type: string
          format: date-time
...
    DeactivateRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 1000
...
security:
- bearerAuth: []